[rename] 뉴스아이템 타입 폴더위치 변경
