+++ src/types/TNews/TNewsItemTypes.ts
